requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
openai==1.3.7
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
itsdangerous==2.1.2
email-validator==2.1.0
Pillow==10.1.0
